Replace duplicate svgs, move some svg to their component's folder
