+++ next/src/assets/images/index.ts
@@ -1,6 +1,5 @@
 export { default as AddressIcon } from './adresa.svg'
 export { default as BrandSVG } from './brand.svg'
-export { default as DaneAPoplatky } from './daneAPoplatky.svg'
 export { default as EmailIcon } from './e-mail.svg'
 export { default as FacebookIcon } from './facebook.svg'
 export { default as InstagramIcon } from './instagram.svg'
