+++ next/src/components/common/HomepageSearch/HomePageSearchField.tsx
@@ -6,7 +6,7 @@ import {
   SetStateAction,
 } from 'react'
 
-import SearchIcon from '@/src/assets/images/search-icon.svg'
+import { SearchIcon } from '@/src/assets/ui-icons'
 import Button from '@/src/components/common/Button/Button'
 
 type HomePageSearchFieldProps = {
