+++ next/src/components/common/OrganizationalStructure_Deprecated/OrganizationalStructureAccordion_Deprecated.tsx
@@ -1,7 +1,6 @@
 import { useToggle } from 'usehooks-ts'
 
-import ChevronDown from '@/src/assets/images/chevron-down-thin.svg'
-import ChevronDownSmall from '@/src/assets/images/chevron-down-thin-small.svg'
+import { ChevronDownIcon } from '@/src/assets/ui-icons'
 import { GetGroupMembersRecursiveResult } from '@/src/services/ms-graph/types'
 import cn from '@/src/utils/cn'
 
@@ -43,8 +42,7 @@ const OrganizationalStructureAccordion = ({
         {/* TODO fix chevron toggles - rotating 180 isn't quite right as the chevron 'jumps' higher - we should use
         different up-down asset */}
         <div className={cn('ml-auto pt-2.5', { 'rotate-180 pt-5': isOpen })}>
-          <ChevronDown className="hidden lg:flex" />
-          <ChevronDownSmall className="flex lg:hidden" />
+          <ChevronDownIcon className="size-6 lg:size-8" />
         </div>
       </div>
